% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condSens.alt.R
\name{condSens.alt}
\alias{condSens.alt}
\title{Sensitivity analysis for conditional exposure effects based on an alternative assumption.}
\usage{
condSens.alt(
  data = NULL,
  outcome = NULL,
  fitmodel,
  model = "lm",
  k,
  p,
  exposure.interest,
  bound,
  delta.range,
  delta.diff = 0.1,
  decimal.p = 3,
  report.result = TRUE,
  only.sig = TRUE,
  n.visual.delta = 3
)
}
\arguments{
\item{data}{A data frame in which contains outcome, measured confounders, exposures. The order of variables (outcome, measured confounders, exposures) that make up the data is important. See Examples.}

\item{outcome}{The name of outcome variable.}

\item{fitmodel}{Model output fitted using one of "lm", "randomforest", and "gbm".}

\item{model}{What model was fitted (one of "lm", "randomforest", and "gbm").}

\item{k}{The number of measured confounders.}

\item{p}{The number of exposures.}

\item{exposure.interest}{Vector of names for exposures of interest.}

\item{bound}{The range of \eqn{(\rho_{1}, \ldots, \rho_{p})}. The order of inputs is (\eqn{\rho_{1,max}, \ldots, \rho_{p,max}, } \eqn{\rho_{1,min}, \ldots, } \eqn{ \rho_{p,min}}). See Examples.}

\item{delta.range}{The range of \eqn{(\delta_{min}, \delta_{max})}. See Examples.}

\item{delta.diff}{The increment of the sequence of \eqn{\delta}. Default: 0.1.}

\item{decimal.p}{Number of digits after the decimal point of numbers appearing in the result table.}

\item{report.result}{Whether to output the result table. Default: TRUE.}

\item{only.sig}{Whether to output results only for significant exposure(s). Default: TRUE.}

\item{n.visual.delta}{The number of \eqn{\delta} in the plot to be printed. This is only used when \code{only.sig=FALSE}. Default: 3.}
}
\value{
The object is a data.frame with the following components:
\itemize{
\item When \code{only.sig=FALSE}:
\itemize{
\item \code{label}: Exposures
\item \code{model_output}: The coefficients of exposures
\item \code{cond_min}: The minimum of conditional single- and joint-exposure effect
\item \code{cond_max}: The maximum of conditional single- and joint-exposure effect
\item \code{delta}: The values of \eqn{\delta}
\item \code{joint}: Whether the value is a conditional joint-exposure effect
}
\item When \code{only.sig=TRUE}:
\itemize{
\item \code{bs}: The coefficient of significant exposure
\item \code{bias_min}: Minimum value of bias
\item \code{bias_max}: Maximum value of bias
\item \code{delta}: The values of \eqn{\delta}
\item \code{cond_min}: The minimum of conditional single-exposure effect
\item \code{cond_max}: The maximum of conditional single-exposure effect
}
}

To generate the plot of results for the \code{condSens.alt}, use the \code{\link{SensPlot}} function.
}
\description{
\code{condSens.alt()} is an alternative sensitivity analysis method using assumption for the conditional distribution of U | (L, X) rather than for the joint distribution of (U, L, X).
}
\details{
See Section 2.4.3 in Jeong et al. (2024) for details.
}
\examples{
# if (!require("MASS", quietly=TRUE)) install.packages("MASS")
library(MASS)

## Our simple running example
set.seed(231113)
n.sample <- 10000
delta <- -0.2
rho1 <- 0.1
rho2 <- 0.2
r11 <- 0.6
dat <- as.data.frame(mvrnorm(n=n.sample, c(0, 0, 0),
                             matrix(c(1,rho1,rho2,rho1,1,r11,rho2,r11,1),3,3),
                             empirical=TRUE))
colnames(dat) <- c("U", "X1", "X2")
dat$Y <- 0.5 * dat$X1 + 0.7 * dat$X2 + delta * dat$U + rnorm(n.sample)

dat_noU <- dat[,c("Y", "X1", "X2")]
fit.model <- lm(Y ~ X1 + X2, data=dat_noU)

rho1_star <- seq(0.05, 0.15, 0.01)
rho2_star <- seq(0.15, 0.25, 0.01)
rho_grid <- expand.grid(rho1_star, rho2_star)

rho_star_f <- function(i, j){
  t(c(i, j)) \%*\% solve(matrix(c(1, 0.6, 0.6, 1), 2, 2))
}
rho_star <- t(sapply(1:nrow(rho_grid), function(x) {
  rho_star_f(rho_grid[x,1], rho_grid[x,2])
}))

rho1_star_min <- min(rho_star[,1])
rho1_star_max <- max(rho_star[,1])
rho2_star_min <- min(rho_star[,2])
rho2_star_max <- max(rho_star[,2])

## All exposures
result.alt <- condSens.alt(data=dat_noU, outcome="Y", fitmodel=fit.model, model="lm",
                           k=0, p=2, exposure.interest=c("X1", "X2"),
                           bound=c(rho1_star_max, rho2_star_max,  # upper bounds
                                   rho1_star_min, rho2_star_min), # lower bounds
                           delta.range=c(-0.3, -0.1), delta.diff=0.1, decimal.p=3,
                           report.result=TRUE, only.sig=FALSE, n.visual.delta=3)
SensPlot(result.alt$result, myxlim=c(0.4, 1.3))

## Only significant exposures
result.alt.sig <- condSens.alt(data=dat_noU, outcome="Y", fitmodel=fit.model, model="lm",
                               k=0, p=2, exposure.interest=c("X1", "X2"),
                               bound=c(rho1_star_max, rho2_star_max,  # upper bounds
                                       rho1_star_min, rho2_star_min), # lower bounds
                               delta.range=c(-0.3, -0.1), delta.diff=0.1, decimal.p=3,
                               report.result=TRUE, only.sig=TRUE, n.visual.delta=3)
SensPlot(result.alt.sig$result$X1)
SensPlot(result.alt.sig$result$X2)

}
\references{
Jeong B, Lee S, Ye S, Lee D, Lee W (2024):
Sensitivity analysis for effects of multiple exposures in the presence of unmeasured confounding
\emph{xxx}. DOI: xxx.
}
\seealso{
\code{\link[SAMU]{SensPlot}}
}
\keyword{Methods}
