% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condSens.R
\name{condSens}
\alias{condSens}
\title{Sensitivity analysis for conditional exposure effects.}
\usage{
condSens(
  data = NULL,
  outcome = NULL,
  fitmodel,
  model = "lm",
  k,
  p,
  bound,
  delta.range,
  delta.diff = 0.1,
  decimal.p = 3,
  report.result = TRUE,
  only.sig = TRUE,
  n.visual.delta = 4
)
}
\arguments{
\item{data}{A data frame in which contains outcome, measured confounders, exposures. The order of variables (outcome, measured confounders, exposures) that make up the data is important.}

\item{outcome}{The name of variable for outcome.}

\item{fitmodel}{Model output fitted using one of "lm", "randomforest", and "gbm".}

\item{model}{What model was fitted (one of "lm", "randomforest", and "gbm").}

\item{k}{The number of measured confounders.}

\item{p}{The number of exposures.}

\item{bound}{The range of \eqn{(\phi_{1}, \ldots, \phi_{k})} and \eqn{(\rho_{1}, \ldots, \rho_{p})}. The order of inputs is c(\eqn{\phi_{1,max}, \ldots, \phi_{k,max}, } \eqn{\rho_{1,max}, \ldots, \rho_{p,max}, \phi_{1,min}, \ldots, \phi_{k,min}, \rho_{1,min}, \ldots, \rho_{p,min}}). See Examples.}

\item{delta.range}{The range of \eqn{(\delta_{min}, \delta_{max})}. See Examples.}

\item{delta.diff}{The increment of the sequence of \eqn{\delta}. Default: 0.1.}

\item{decimal.p}{Number of digits after the decimal point of numbers appearing in the result table.}

\item{report.result}{Whether to output the result table. Default: TRUE.}

\item{only.sig}{Whether to output results only for significant variables. Default: TRUE.}

\item{n.visual.delta}{The Number of \eqn{\delta} in the plot to be printed. This is only used when only.sig=FALSE. Default: 4.}
}
\value{
An object of class \code{condSens}. The object is a data.frame with the following components: 1. only.sig=FALSE
\item{label}{Exposures}
\item{model_output}{The coefficients of exposures}
\item{cond_min}{The minimum of conditional single- and joint-exposure effect}
\item{cond_max}{The maximum of conditional single- and joint-exposure effect}
\item{delta}{The values of \eqn{\delta}}
\item{joint}{Whether the value is a conditional joint-exposure effect}
\item{bs}{The coefficient of significant exposure}
\item{bias_min}{Minimum value of bias}
\item{bias_max}{Maximum value of bias}
\item{delta}{The values of \eqn{\delta}}
\item{cond_min}{The minimum of conditional single-exposure effect}
\item{cond_max}{The maximum of conditional single-exposure effect}

To generate the plot of results for the \code{condSens}, use the \code{\link{SensPlot}} functions.
}
\description{
\code{condSens()} is the main function of condSens and
performs the sensitivity analysis for estimating the sensitivity range(s).
}
\details{
See Jeong et al. (2024) for details.
}
\examples{
## Import data in out paper and NA omit
library(readr)
ul="https://raw.githubusercontent.com/lizzyagibson/SHARP.Mixtures.Workshop/master/Data/studypop.csv"
data <- as.data.frame(read_csv(url(ul))); data <- na.omit(data)

## Log transformation for outcome
data$TELOMEAN <- log(data$TELOMEAN)

## Change the order of variables:: 1 outcome - 11 confounders - 18 exposures
data <- data[,c("TELOMEAN",
                # confounders
                "bmi_cat3","edu_cat","race_cat","male","ln_lbxcot","age_cent",
                "LBXWBCSI","LBXLYPCT","LBXMOPCT","LBXEOPCT","LBXBAPCT",
                # exposures
                "LBX074LA","LBX099LA","LBX138LA","LBX153LA","LBX170LA","LBX180LA",
                "LBX187LA","LBX194LA","LBXPCBLA","LBXHXCLA","LBX118LA",
                "LBXD03LA","LBXD05LA","LBXD07LA",
                "LBXF03LA","LBXF04LA","LBXF05LA","LBXF08LA"
)]
colnames(data) <- c("TELOMEAN",
                    "bmi_cat3","edu_cat","race_cat","male","ln_lbxcot","age_cent",
                    "LBXWBCSI","LBXLYPCT","LBXMOPCT","LBXEOPCT","LBXBAPCT",
                    "PCB74","PCB99","PCB138","PCB153","PCB170","PCB180",
                    "PCB187","PCB194","PCB126","PCB169","PCB118",
                    "Dioxin1","Dioxin2","Dioxin3",
                    "Furan1","Furan2","Furan3","Furan4")

## Change bmi (binary) 0:< 30 1: >=30
data$bmi_cat3 <- ifelse(data$bmi_cat3 >= 3, 1, 0)

## Log transformation of exposures
names.logtrans <- c("PCB74","PCB99","PCB138","PCB153","PCB170","PCB180",
                    "PCB187","PCB194","PCB126","PCB169","PCB118",
                    "Dioxin1","Dioxin2","Dioxin3",
                    "Furan1","Furan2","Furan3","Furan4")
data[,names.logtrans] <- log(data[,names.logtrans])

## Standardized variables
data_r3 <- data
data_r3[,-c(1)] <- scale(data_r3[,-c(1)])

# Sensitivity analysis for conditional exposure effect
## Working model: Linear regression (confounders adjusted model)
## Number of confounders
k <- 11
## Number of exposures
p <- 18

## Fitting the linear model
fitmodel <- lm(TELOMEAN ~., data=data_r3)

## Sensitivity analysis results
### Figure 6
## L-U is uncorrelated, and X-U is correlated with ranges of (0.18, 95).
## 1. All exposures
rst1 <- condSens(data=data_r3, outcome="TELOMEAN", fitmodel=fitmodel, model="lm",
                 k=k, p=p,
                 bound=c(rep(0, 11), rep(0.95, 18),
                         rep(0, 11), rep(0.18, 18)),
                 delta.range=c(-0.03, -0.01), delta.diff=0.01, decimal.p=3,
                 report.result=TRUE, only.sig=FALSE, n.visual.delta=3)

## 2. Only significant exposure
rst1_sig <- condSens(data=data_r3, outcome="TELOMEAN", fitmodel=fitmodel, model="lm",
                     k=k, p=p,
                     bound=c(rep(0, 11), rep(0.95, 18),
                             rep(0, 11), rep(0.18, 18)),
                     delta.range=c(-0.03, -0.01), delta.diff=0.01, decimal.p=3,
                     report.result=TRUE, only.sig=TRUE, n.visual.delta=3)

}
\references{
Jeong B, Lee S, Ye S, Lee D, Lee W (2024):
Sensitivity analysis for effects of multiple exposures in the presence of unmeasured confounding
\emph{xxx}. DOI: xxx.
}
\seealso{
\code{\link[SAMU]{SensPlot}}
}
\keyword{methods}
