% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarginalSens.R
\name{MarginalSens}
\alias{MarginalSens}
\title{Sensitivity analysis for marginal exposure effects.}
\usage{
MarginalSens(
  data,
  fittedmodel,
  k,
  p,
  delta,
  bound,
  n.col,
  xlim,
  ylim,
  text.which,
  text.size,
  line.size
)
}
\arguments{
\item{data}{A data frame in which contains outcome, measured confounders, exposures. The order of variables (outcome, measured confounders, exposures) that make up the data is important.}

\item{fittedmodel}{Model output fitted using one of "lm", "randomforest", and "gbm".}

\item{k}{The number of measured confounders.}

\item{p}{The number of exposures.}

\item{delta}{The value of \eqn{\delta}.}

\item{bound}{The range of \eqn{(\phi_{1}, \ldots, \phi_{k})} and \eqn{(\rho_{1}, \ldots, \rho_{p})}. The order of inputs is c(\eqn{\phi_{1,max}, \ldots, \phi_{k,max}, } \eqn{\rho_{1,max}, \ldots, \rho_{p,max}, \phi_{1,min}, \ldots, \phi_{k,min}, \rho_{1,min}, \ldots, \rho_{p,min}}). See Examples.}

\item{n.col}{Number of columns to display plot}

\item{xlim}{x-axis range}

\item{ylim}{y-axis range}

\item{text.which}{Coordinates to display slopes}

\item{text.size}{Size of text displaying slopes}

\item{line.size}{Line size of slopes}
}
\description{
Sensitivity analysis for marginal exposure effects.
}
\examples{
\dontrun{
## Import data and NA omit
library (readr)
ul="https://raw.githubusercontent.com/lizzyagibson/SHARP.Mixtures.Workshop/master/Data/studypop.csv"
data <- as.data.frame(read_csv(url(ul))); data <- na.omit(data)

## Log transformation for outcome
data$TELOMEAN <- log(data$TELOMEAN)

## Change the order of variables:: 1 outcome - 11 confounders - 18 exposures
data <- data[,c("TELOMEAN",
                # confounders
                "bmi_cat3","edu_cat","race_cat","male","ln_lbxcot","age_cent",
                "LBXWBCSI","LBXLYPCT","LBXMOPCT","LBXEOPCT","LBXBAPCT",
                # exposures
                "LBX074LA","LBX099LA","LBX138LA","LBX153LA","LBX170LA","LBX180LA",
                "LBX187LA","LBX194LA","LBXPCBLA","LBXHXCLA","LBX118LA",
                "LBXD03LA","LBXD05LA","LBXD07LA",
                "LBXF03LA","LBXF04LA","LBXF05LA","LBXF08LA"
)]
colnames(data) <- c("TELOMEAN",
                    "bmi_cat3","edu_cat","race_cat","male","ln_lbxcot","age_cent",
                    "LBXWBCSI","LBXLYPCT","LBXMOPCT","LBXEOPCT","LBXBAPCT",
                    "PCB74","PCB99","PCB138","PCB153","PCB170","PCB180",
                    "PCB187","PCB194","PCB126","PCB169","PCB118",
                    "Dioxin1","Dioxin2","Dioxin3",
                    "Furan1","Furan2","Furan3","Furan4")

## Change bmi (binary) 0:< 30 1: >=30
data$bmi_cat3 <- ifelse(data$bmi_cat3 >= 3, 1, 0)

## Log transformation of exposures
names.logtrans <- c("PCB74","PCB99","PCB138","PCB153","PCB170","PCB180",
                    "PCB187","PCB194","PCB126","PCB169","PCB118",
                    "Dioxin1","Dioxin2","Dioxin3",
                    "Furan1","Furan2","Furan3","Furan4")
data[,names.logtrans] <- log(data[,names.logtrans])

## Standardized variables
data_r3 <- data
data_r3[,-c(1)] <- scale(data_r3[,-c(1)])

# Sensitivity analysis for conditional exposure effect
## Working model: Linear regression (confounders adjusted model)
## Number of confounders
k <- 11
## Number of exposures
p <- 18

# Sensitivity analysis for marginal exposure effect
## Working model: Random forest (confounders adjusted model)
## 10-fold cross-validation for tuning parameter (mtry)
set.seed(231111)
library(caret)
library(randomForest)
metric <- "RMSE"
control <- trainControl(method="cv", number=10, search="grid")
tunegrid <- expand.grid(.mtry=seq(1, 29, 2))
rf_gridsearch <- train(TELOMEAN~., data=data_r3, method="rf",
                       metric=metric, tuneGrid=tunegrid, trControl=control)

## model fitting
mtry <- as.numeric(rf_gridsearch$bestTune)
rfmodel <- randomForest(TELOMEAN~., data=data_r3, mtry=mtry)

### PDP-based sensitivity analysis results for marginal exposure effect
### Supplementary Figure 22
## L-U is uncorrelated, and X-U correlated with ranges of (0.18, 0.95).
mrst1 <- MarginalSens(data_r3, rfmodel, k=k, p=p, delta=-0.02,
                      bound=c(rep(0, 11), rep(0.95, 18),
                              rep(0, 11), rep(0.18, 18)),
                      n.col=6, xlim=c(-4, 4), ylim=c(-0.2, 0.2),
                      text.which=c(-1.0, 4), text.size=2, line.size=0.8)
}

}
\references{
Jeong B, Lee S, Ye S, Lee D, Lee W (2024):
Sensitivity analysis for effects of multiple exposures in the presence of unmeasured confounding
\emph{xxx}. DOI: xxx.
}
\keyword{methods}
