% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAMU-package.R
\docType{package}
\name{SAMU-package}
\alias{SAMU}
\alias{SAMU-package}
\title{SAMU: Sensitivity analysis for effects of multiple exposures in the presence of unmeasured confounding}
\description{
R package SAMU is sensitivity analysis package for the conditional and marginal exposure effects to possible unmeasured confounding.
}
\references{
Jeong B, Lee S, Ye S, Lee D, Lee W (2024):
Sensitivity analysis for effects of multiple exposures in the presence of unmeasured confounding
\emph{xxx}. DOI: xxx.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sites.google.com/view/lwj221/research/softwares}
}

}
\author{
\strong{Maintainer}: Seungjae Lee \email{seungjae2525@gmail.com}

Authors:
\itemize{
  \item Boram Jeong
  \item Shinhee Ye
  \item Donghwan Lee
  \item Woojoo Lee
}

}
