% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SensPlot.R
\name{SensPlot}
\alias{SensPlot}
\title{Plot for sensitivity analysis results of conditional exposure effect}
\usage{
SensPlot(
  condSens.result,
  title = "Sensitivity analysis",
  myxlim = NULL,
  exposure.n = NULL,
  alpha.range = 0.4,
  alpha.ci = 0.9,
  ytickdiff = 0.01,
  point.size = 1.4,
  h.width = 1,
  axis.title.size = 15,
  axis.text.size = 12,
  label.size = 4
)
}
\arguments{
\item{condSens.result}{Object from \code{condSens()} or \code{condSens.alt()}.}

\item{title}{Plot title name. This is only used when \code{only.sig=FALSE} in \code{condSens()} or \code{condSens.alt()}. Default: "Sensitivity analysis".}

\item{myxlim}{The minimum and maximum values of x-axis. This is only used when \code{only.sig=FALSE} in \code{condSens()} or \code{condSens.alt()}.}

\item{exposure.n}{The variable name(s) of significant exposure(s). This is only used when \code{only.sig=TRUE} in \code{condSens()} or \code{condSens.alt()}. Default: NULL.}

\item{alpha.range}{A opacity of sensitivity interval. Values of \code{alpha.range} range from 0 to 1, with lower values corresponding to more transparent colors. This is only used when \code{only.sig=TRUE} in \code{condSens()} or \code{condSens.alt()}. Default: 0.4.}

\item{alpha.ci}{A opacity of confidence interval. Values of \code{alpha.ci} range from 0 to 1, with lower values corresponding to more transparent colors. This is only used when \code{only.sig=TRUE} in \code{condSens()} or \code{condSens.alt()}. Default: 0.9.}

\item{ytickdiff}{Distance between y-axis tick. This is only used when \code{only.sig=TRUE} in \code{condSens()} or \code{condSens.alt()}. Default: 0.01.}

\item{point.size}{Point size of the the lower and upper bound of the sensitivity interval. This is only used when \code{only.sig=TRUE} in \code{condSens()} or \code{condSens.alt()}. Default: 1.4.}

\item{h.width}{Width of horizon line which represents the effect size. This is only used when \code{only.sig=TRUE} in \code{condSens()} or \code{condSens.alt()}. Default: 1.}

\item{axis.title.size}{Size of x and y axis title. This is only used when \code{only.sig=TRUE} in \code{condSens()} or \code{condSens.alt()}. Default: 15.}

\item{axis.text.size}{Size of x and y axis text. This is only used when \code{only.sig=TRUE} in \code{condSens()} or \code{condSens.alt()}. Default: 12.}

\item{label.size}{label text size. This is only used when \code{only.sig=TRUE} in \code{condSens()} or \code{condSens.alt()}. Default: 4.}
}
\description{
Plot for sensitivity analysis results of conditional exposure effect.
When \code{only.sig=FALSE} in \code{condSens()} or \code{condSens.alt()}, the resulting plot for all exposures appears.
When \code{only.sig=TRUE} in \code{condSens()} or \code{condSens.alt()}, the resulting plot(s) for only significant exposure(s) appear.
}
\examples{
## Import data in out paper and NA omit
library(readr)
ul="https://raw.githubusercontent.com/lizzyagibson/SHARP.Mixtures.Workshop/master/Data/studypop.csv"
data <- as.data.frame(read_csv(url(ul))); data <- na.omit(data)

## Log transformation for outcome
data$TELOMEAN <- log(data$TELOMEAN)

## Change the order of variables:: 1 outcome - 11 confounders - 18 exposures
data <- data[,c("TELOMEAN",
                # confounders
                "bmi_cat3","edu_cat","race_cat","male","ln_lbxcot","age_cent",
                "LBXWBCSI","LBXLYPCT","LBXMOPCT","LBXEOPCT","LBXBAPCT",
                # exposures
                "LBX074LA","LBX099LA","LBX138LA","LBX153LA","LBX170LA","LBX180LA",
                "LBX187LA","LBX194LA","LBXPCBLA","LBXHXCLA","LBX118LA",
                "LBXD03LA","LBXD05LA","LBXD07LA",
                "LBXF03LA","LBXF04LA","LBXF05LA","LBXF08LA"
)]
colnames(data) <- c("TELOMEAN",
                    "bmi_cat3","edu_cat","race_cat","male","ln_lbxcot","age_cent",
                    "LBXWBCSI","LBXLYPCT","LBXMOPCT","LBXEOPCT","LBXBAPCT",
                    "PCB74","PCB99","PCB138","PCB153","PCB170","PCB180",
                    "PCB187","PCB194","PCB126","PCB169","PCB118",
                    "Dioxin1","Dioxin2","Dioxin3",
                    "Furan1","Furan2","Furan3","Furan4")

## Change bmi to binary // 0: < 30 1: >= 30
data$bmi_cat3 <- ifelse(data$bmi_cat3 >= 3, 1, 0)

## Change race to binary // 4 (Non-Hispanic White) vs others
data$race_cat <- ifelse(data$race_cat == 4, 1, 0)

## Change education to binary // after college (3 and 4) vs before college (1 and 2)
data$edu_cat <- ifelse(data$edu_cat \%in\% c(3,4), 1, 0)

## Log transformation of exposures
names.logtrans <- c("PCB74","PCB99","PCB138","PCB153","PCB170","PCB180",
                    "PCB187","PCB194","PCB126","PCB169","PCB118",
                    "Dioxin1","Dioxin2","Dioxin3",
                    "Furan1","Furan2","Furan3","Furan4")
data[,names.logtrans] <- log(data[,names.logtrans])

## Standardized variables
data_r3 <- data
data_r3[,-c(1)] <- scale(data_r3[,-c(1)])

# Sensitivity analysis for conditional exposure effect
## Working model: Linear regression (confounders adjusted model)
## Number of confounders
k <- 11
## Number of exposures
p <- 18

## Fitting the linear model
fitmodel <- lm(TELOMEAN ~., data=data_r3)

## Sensitivity analysis results
### For Figure 6 (a)
## L-U is uncorrelated, and X-U is correlated with ranges of (0.18, 95).
## 1. All exposures
rst1 <- condSens(data=data_r3, outcome="TELOMEAN", fitmodel=fitmodel, model="lm",
                 k=k, p=p,
                 bound=c(rep(0, 11), rep(0.95, 18),  # upper bounds
                         rep(0, 11), rep(0.18, 18)), # lower bounds
                 delta.range=c(-0.03, -0.01), delta.diff=0.01, decimal.p=3,
                 report.result=TRUE, only.sig=FALSE, n.visual.delta=3)
SensPlot(condSens.result=rst1$result, myxlim=c(-0.2, 0.2),
         title="Sensitivity analysis\nCase 1")

## 2. Only significant exposure
### For Figure 6 (b)
rst1_sig <- condSens(data=data_r3, outcome="TELOMEAN", fitmodel=fitmodel, model="lm",
                     k=k, p=p,
                     bound=c(rep(0, 11), rep(0.95, 18),  # upper bounds
                             rep(0, 11), rep(0.18, 18)), # lower bounds
                     delta.range=c(-0.03, -0.01), delta.diff=0.01, decimal.p=3,
                     report.result=TRUE, only.sig=TRUE, n.visual.delta=3)
SensPlot(condSens.result=rst1_sig$result$Furan1, exposure.n="Furan1",
         alpha.range=0.4, alpha.ci=0.9, ytickdiff=0.01, point.size=1.4, h.width=1,
         axis.title.size=13, axis.text.size=10, label.size=3.5)

}
\references{
Jeong B, Lee S, Ye S, Lee D, Lee W (2024):
Sensitivity analysis for effects of multiple exposures in the presence of unmeasured confounding
\emph{xxx}. DOI: xxx.
}
\seealso{
\code{\link[SAMU]{condSens}} and \code{\link[SAMU]{condSens.alt}}
}
\keyword{Plot}
